+++ wai-csrf/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Depend on `ram` instead of `memory`.
 
+* Use `__Host-CSRF-TOKEN` as default cookie name.
+
 
 # Version 0.1